Split views into views.py
